+++ CHANGELOG.md
@@ -21,7 +21,7 @@ All notable changes to this project will be documented in this file. This change
 * `cohens-u1-normal`, `cohens-u2-normal`, `cohens-u3-normal` - original (parametric) version of Cohen's U statistics
 * `cohens-u1` - based on Cohen's formula and `cohens-u2` value.
 * `cir`, `pava` - isotonic regression + related interpolators.
-* `ks-test-two-samples` - added jittering the data and sides calculation for `:exact` method
+* `ks-test-two-samples` - added jittering the data, ties solving and sides calculation for `:exact` method 
 
 ### Fixed
 
@@ -36,6 +36,7 @@ All notable changes to this project will be documented in this file. This change
 ### Changed
 
 * [breaking] `yeo-johnson-transformation` accepts a map as last argument (was `alpha`)
+* `ks-test-two-samples` - `:distinct?` now is set to `:ties`
 
 ## [3.0.0 alpha3]